requirements.txt
panda
Pillow
numpy
scipy
matplotlib
dash
dash_core_components
dash_html_components
plotly
dash_renderer